% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_stratified_analytical_app.R
\name{additive.power.analytical.app}
\alias{additive.power.analytical.app}
\title{Analyze Power for a Stratified Additive RMST Model (Analytic)}
\usage{
additive.power.analytical.app(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  strata_var,
  sample_sizes,
  linear_terms = NULL,
  L,
  alpha = 0.05
)
}
\arguments{
\item{sample_sizes}{A numeric vector of sample sizes *per stratum* to calculate power for.}

\item{alpha}{The significance level (Type I error rate).}
}
\value{
A list containing the results.
}
\description{
Performs power analysis for a stratified, additive RMST model.
}
\keyword{internal}
