% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_gam_boot_app.R
\name{additive.power.boot.app}
\alias{additive.power.boot.app}
\title{Calculate Power for a Semiparametric Additive RMST Model via Simulation}
\usage{
additive.power.boot.app(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  strata_var = NULL,
  sample_sizes,
  linear_terms = NULL,
  smooth_terms = NULL,
  L,
  n_sim = 1000,
  alpha = 0.05,
  parallel.cores = 1
)
}
\arguments{
\item{sample_sizes}{A numeric vector of sample sizes per arm/stratum.}
}
\value{
A list containing results.
}
\description{
Performs a power analysis for given sample sizes using a flexible,
  semiparametric additive model for the RMST based on pseudo-observations.
}
\keyword{internal}
