% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_app_method_datasets.R
\docType{data}
\name{app_method_datasets}
\alias{app_method_datasets}
\title{App Method Datasets}
\format{
A named list of data frames. Names follow
`"<method_stub>_<dataset_type>"`, for example
`linear_ipcw_model_complete` and `linear_ipcw_model_missing`.
}
\usage{
app_method_datasets
}
\description{
Named list of app-ready datasets generated by
\code{\link{generate_app_method_datasets}}.
}
\keyword{datasets}
