% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ipcw_analytical_app.R
\name{.estimate_linear_ipcw_params}
\alias{.estimate_linear_ipcw_params}
\title{Linear IPCW RMST Model Power and Sample Size Calculation (Analytic)}
\usage{
.estimate_linear_ipcw_params(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  linear_terms,
  L
)
}
\description{
An internal function for the Shiny app.
}
\keyword{internal}
