% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicative_stratified_analytical_app.R
\name{.estimate_multiplicative_stratified_params}
\alias{.estimate_multiplicative_stratified_params}
\title{Internal Helper to Estimate Multiplicative Stratified Model Parameters}
\usage{
.estimate_multiplicative_stratified_params(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  strata_var,
  linear_terms,
  L
)
}
\value{
A list containing `beta_effect`, `se_beta_n1`, and `n_strata`.
}
\description{
This internal function contains the common logic for estimating the
  log-RMST ratio and its variance from pilot data using a log-linear approximation.
}
\keyword{internal}
