% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_gam_boot_app.R
\name{.get_gam_simulation_runner}
\alias{.get_gam_simulation_runner}
\title{Internal Factory for Additive GAM RMST Simulation}
\usage{
.get_gam_simulation_runner(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  strata_var,
  linear_terms,
  smooth_terms,
  L,
  alpha,
  n_sim,
  parallel.cores
)
}
\value{
A function that takes `n_per_group` and runs the simulation.
}
\description{
This internal function prepares and returns a configured simulation function.
}
\keyword{internal}
