% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ipcw_boot_app.R
\name{.get_linear_ipcw_simulation_runner}
\alias{.get_linear_ipcw_simulation_runner}
\title{Internal Factory for Linear IPCW RMST Simulation}
\usage{
.get_linear_ipcw_simulation_runner(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  linear_terms,
  L,
  alpha,
  n_sim,
  parallel.cores
)
}
\value{
A function that takes `n_per_arm` and runs the simulation.
}
\description{
This internal function prepares and returns a configured simulation function.
}
\keyword{internal}
