% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ipcw_analytical_app.R
\name{linear.power.analytical.app}
\alias{linear.power.analytical.app}
\title{Analyze Power for a Linear RMST Model (Analytic)}
\usage{
linear.power.analytical.app(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  sample_sizes,
  linear_terms = NULL,
  L,
  alpha = 0.05
)
}
\arguments{
\item{sample_sizes}{A numeric vector of sample sizes *per arm* to calculate power for.}

\item{alpha}{The significance level for the power calculation (Type I error rate).}
}
\value{
A `list` containing results.
}
\description{
Performs power analysis using a direct formula based on the
  asymptotic variance estimator for the linear RMST model.
}
\keyword{internal}
