% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ipcw_boot_app.R
\name{linear.power.boot.app}
\alias{linear.power.boot.app}
\title{Analyze Power for a Linear RMST Model via Simulation}
\usage{
linear.power.boot.app(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  sample_sizes,
  linear_terms = NULL,
  L,
  n_sim = 1000,
  alpha = 0.05,
  parallel.cores
)
}
\arguments{
\item{sample_sizes}{A numeric vector of sample sizes *per arm* to calculate power for.}
}
\value{
A `list` containing results.
}
\description{
Performs a power analysis for given sample sizes based on the direct
  linear regression model for RMST, using a bootstrap simulation approach.
}
\keyword{internal}
