% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ipcw_analytical_app.R
\name{linear.ss.analytical.app}
\alias{linear.ss.analytical.app}
\title{Find Sample Size for a Linear RMST Model (Analytic)}
\usage{
linear.ss.analytical.app(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  target_power,
  linear_terms = NULL,
  L,
  alpha = 0.05,
  n_start = 50,
  n_step = 25,
  max_n_per_arm = 2000
)
}
\arguments{
\item{target_power}{A single numeric value for the desired power (e.g., 0.80 or 0.90).}

\item{alpha}{The significance level (Type I error rate).}

\item{n_start}{The starting sample size *per arm* for the search.}

\item{n_step}{The increment in sample size at each step of the search.}

\item{max_n_per_arm}{The maximum sample size *per arm* to search up to.}
}
\value{
A `list` containing results.
}
\description{
Calculates the required sample size for a target power using an
  analytic formula based on the methods of Tian et al. (2014).
}
\keyword{internal}
