% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ipcw_boot_app.R
\name{linear.ss.boot.app}
\alias{linear.ss.boot.app}
\title{Find Sample Size for a Linear RMST Model via Simulation}
\usage{
linear.ss.boot.app(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  target_power,
  linear_terms = NULL,
  L,
  n_sim = 1000,
  alpha = 0.05,
  patience = 5,
  n_start = 50,
  n_step = 25,
  max_n_per_arm = 2000,
  parallel.cores
)
}
\arguments{
\item{target_power}{A single numeric value for the target power (e.g., 0.80).}

\item{patience}{The number of consecutive non-improving steps before terminating.}

\item{n_start}{The starting sample size *per arm* for the search.}

\item{n_step}{The increment in sample size at each step of the search.}

\item{max_n_per_arm}{The maximum sample size *per arm* to search up to.}
}
\value{
A `list` containing results.
}
\description{
Performs an iterative sample size search to achieve a target power.
}
\keyword{internal}
