% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicative_stratified_boot_app.R
\name{MS.power.boot.app}
\alias{MS.power.boot.app}
\title{Analyze Power for a Multiplicative Stratified RMST Model via Simulation}
\usage{
MS.power.boot.app(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  strata_var,
  sample_sizes,
  linear_terms = NULL,
  L,
  n_sim = 1000,
  alpha = 0.05,
  parallel.cores = 1
)
}
\arguments{
\item{time_var}{A character string for the time-to-event variable.}

\item{status_var}{A character string for the event status variable.}

\item{arm_var}{A character string for the treatment arm variable.}

\item{strata_var}{A character string for the stratification variable.}

\item{sample_sizes}{A numeric vector of sample sizes *per stratum* to calculate power for.}

\item{linear_terms}{Optional character vector of covariates for the model.}

\item{L}{The numeric truncation time for RMST.}

\item{n_sim}{Number of bootstrap simulations.}

\item{alpha}{The significance level.}

\item{parallel.cores}{Number of cores for parallel processing.}
}
\description{
Performs power analysis based on a multiplicative model for RMST
  for stratified trials, using a bootstrap simulation approach.
  #' @param pilot_data A `data.frame` with pilot study data.
}
\keyword{internal}
