% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicative_stratified_boot_app.R
\name{MS.ss.boot.app}
\alias{MS.ss.boot.app}
\title{Estimate Sample Size for a Multiplicative Stratified RMST Model via Simulation}
\usage{
MS.ss.boot.app(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  strata_var,
  target_power,
  linear_terms = NULL,
  L,
  n_sim = 1000,
  alpha = 0.05,
  parallel.cores = 1,
  patience = 5,
  n_start = 50,
  n_step = 25,
  max_n_per_arm = 2000
)
}
\arguments{
\item{pilot_data}{A `data.frame` with pilot study data.}

\item{time_var}{A character string for the time-to-event variable.}

\item{status_var}{A character string for the event status variable.}

\item{arm_var}{A character string for the treatment arm variable.}

\item{strata_var}{A character string for the stratification variable.}

\item{target_power}{A single numeric value for the target power (e.g., 0.80).}

\item{linear_terms}{Optional vector of covariates for the model.}

\item{L}{The numeric truncation time for RMST.}

\item{n_sim}{Number of bootstrap simulations per search step.}

\item{alpha}{The significance level.}

\item{parallel.cores}{Number of cores for parallel processing.}

\item{patience}{Number of consecutive non-improving steps in the search before terminating.}

\item{n_start}{Starting sample size per stratum for the search.}

\item{n_step}{Increment for the sample size search.}

\item{max_n_per_arm}{Maximum sample size per stratum to try.}
}
\description{
Performs sample size estimation based on a multiplicative model for RMST
  for stratified trials, using iterative bootstrap simulations.
}
\keyword{internal}
