% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_imports.R
\docType{package}
\name{RMSTpowerBoostApp-package}
\alias{RMSTpowerBoostApp}
\alias{RMSTpowerBoostApp-package}
\title{Internal imports for the Shiny app package}
\description{
Shiny application and analysis utilities for RMST power and sample size calculations.
}
\author{
\strong{Maintainer}: Arnab Aich \email{118152677+arnabaich96@users.noreply.github.com}

}
\keyword{internal}
